highlight_text/requirements.txt
boto3
Pillow
regex
